requirements.txt
streamli
pandas
seaborn
matplotlib.pyplot